chore: move graph to package
